Admin panel aesthetics (WIP)
Admins will be able to invite users (which immediately registers them and emails a password), or give them a code to register normally with. Admins can delete users but most importantly can create rooms. They create rooms with names and can modify it's template so standard teachers can just pick the room and assign students where they want. Admins can also be in control of users for the whole organisation